refactor: move browser implementation and shared code to their own package
