+++ packages/ts-endpoint-browser/README.md
@@ -1,0 +1,11 @@
+[![lerna](https://img.shields.io/badge/maintained%20with-lerna-cc00ff.svg)](https://lerna.js.org/)
+
+# ts-endpoint-browser
+
+### description
+
+`ts-endpoint-browser` uses endpoints created with `ts-endpoint` to generate a fetch client to use in browsers.
+
+### docs
+documentation of every package can be found at https://unruffled-fermat-d9a516.netlify.app
+
