+++ packages/ts-endpoint-browser/jest.config.js
@@ -1,0 +1,7 @@
+const base = require('../../jest.config.base.js');
+
+module.exports = {
+  ...base,
+  name: 'ts-endpoint-browser',
+  displayName: 'ts-endpoint-browser',
+};
