+++ packages/ts-endpoint-browser/jest.typetests-config.js
@@ -1,0 +1,7 @@
+const base = require('../../jest.typetests-config.base.js');
+
+module.exports = {
+  ...base,
+  name: 'ts-endpoint-browser',
+  displayName: 'ts-endpoint-browser',
+};
