+++ packages/ts-endpoint-browser/src/__test__/fetch.spec.ts
@@ -1,10 +1,10 @@
 import 'isomorphic-fetch';
 import { GetFetchHTTPClient } from '../fetch';
-import { Endpoint } from '../../Endpoint';
+import { Endpoint } from 'ts-endpoint';
 import { StaticHTTPClientConfig } from '../config';
 import * as t from 'io-ts';
 import { isLeft } from 'fp-ts/lib/Either';
-import { DecodeErrorStatus, NetworkErrorStatus } from '../../shared/errors';
+import { DecodeErrorStatus, NetworkErrorStatus } from 'ts-shared/src/errors';
 
 const options: StaticHTTPClientConfig = {
   protocol: 'http',
