+++ packages/ts-endpoint-browser/src/__test__/swr.spec.ts
@@ -1,9 +1,9 @@
 import 'isomorphic-fetch';
-import { Endpoint } from '../../Endpoint';
+import { Endpoint } from 'ts-endpoint';
 // import { renderHook } from '@testing-library/react-hooks';
 import { StaticHTTPClientConfig } from '../config';
 import * as t from 'io-ts';
-import { DecodeErrorStatus } from '../../shared/errors';
+import { DecodeErrorStatus } from 'ts-shared/src/errors';
 import { GetSWRImpl } from '../swr';
 import { cache } from 'swr';
 import { isLeft } from 'fp-ts/lib/Either';
