+++ packages/ts-endpoint-browser/tsconfig.json
@@ -2,7 +2,7 @@
   "extends": "../../tsconfig.json",
   "compilerOptions": {
     "target": "es5",
-    "lib": ["esnext"],
+    "lib": ["dom", "esnext"],
     "outDir": "./lib",
     "esModuleInterop": true
   },
