+++ packages/ts-endpoint-express/README.md
@@ -1,0 +1,11 @@
+[![lerna](https://img.shields.io/badge/maintained%20with-lerna-cc00ff.svg)](https://lerna.js.org/)
+
+# ts-endpoint-express
+
+### description
+
+`ts-endpoint-express` uses endpoints created with `ts-endpoint` to add routes to an `express` router in a typesafe manner.
+
+### docs
+documentation of every package can be found at https://unruffled-fermat-d9a516.netlify.app
+
