+++ packages/ts-endpoint/README.md
@@ -1,0 +1,12 @@
+[![lerna](https://img.shields.io/badge/maintained%20with-lerna-cc00ff.svg)](https://lerna.js.org/)
+
+# ts-endpoint
+
+### docs
+documentation of every package can be found at https://unruffled-fermat-d9a516.netlify.app
+
+### examples
+a working monorepo example can be found under the `/examples` foder.
+
+### contibuting
+just pick a card on the github projects found here, please be sure to follow `conventional-commits` guidelines.
