+++ packages/ts-endpoint/tsconfig.json
@@ -2,7 +2,7 @@
   "extends": "../../tsconfig.json",
   "compilerOptions": {
     "target": "es5",
-    "lib": ["dom", "esnext"],
+    "lib": ["esnext"],
     "outDir": "./lib",
     "esModuleInterop": true
   },
