+++ packages/ts-shared/jest.config.js
@@ -1,0 +1,7 @@
+const base = require('../../jest.config.base.js');
+
+module.exports = {
+  ...base,
+  name: 'shared',
+  displayName: 'shared',
+};
