+++ packages/ts-shared/package.json
@@ -1,14 +1,15 @@
 {
-  "name": "ts-endpoint-express",
-  "version": "1.1.0-alpha.12",
-  "description": "Adds typesafe endpoints to your express app.",
+  "name": "ts-shared",
+  "version": "1.0.0",
+  "private": "true",
+  "description": "shared lib for ts-endpoint packages.",
   "files": [
     "lib"
   ],
   "main": "lib/index.js",
-  "repository": "https://github.com/fes300/ts-endpoint/packages/ts-endpoint-express",
+  "repository": "https://github.com/fes300/ts-endpoint/packages/shared",
   "bugs": {
-    "url": "https://github.com/fes300/ts-endpoint/issues?q=is%3Aopen+is%3Aissue+project%3Afes300%2Fopen-source%2F4"
+    "url": "https://github.com/fes300/ts-endpoint/issues"
   },
   "author": "Federico Sordillo <federicosordillo@gmail.com>",
   "license": "ISC",
@@ -17,27 +18,28 @@
   },
   "scripts": {
     "build": "rm -rf ./lib && tsc",
-    "test": "jest --config=jest.typetests-config.js && jest --config=jest.config.js"
+    "test": "jest --config=jest.typetests-config.js --passWithNoTests && jest --config=jest.config.js"
   },
   "peerDependencies": {
-    "express": "^4.17.1",
     "fp-ts": "^2.0.0",
     "io-ts": "^2.2.7"
   },
   "devDependencies": {
     "@types/express": "^4.17.7",
     "@types/jest": "^26.0.7",
+    "@types/qs": "^6.9.3",
     "dts-jest": "^23.3.0",
-    "express": "^4.17.1",
     "jest": "^26.4.2",
     "ts-jest": "^26.1.1",
-    "typescript": "^4.0.2"
+    "typescript": "4.0.2"
   },
   "dependencies": {
-    "ts-endpoint": "^1.5.0-alpha.14",
+    "io-ts": "^2.2.7",
     "io-ts-types": "^0.5.8",
     "monocle-ts": "^2.3.3",
     "newtype-ts": "^0.3.4",
+    "qs": "^6.9.4",
+    "rxjs": "^6.6.0",
     "typelevel-ts": "^0.4.0"
   }
 }
