+++ packages/ts-shared/src/BearerToken.ts
