+++ packages/ts-shared/src/Percentage.ts
