+++ packages/ts-shared/src/__test__/BearerToken.spec.ts
