+++ packages/ts-shared/src/errors.ts
